Switch to long long for get_sys_mem()
We might as well as a future proofing.

+++ include/Data_Functions.hpp
@@ -11,7 +11,7 @@
 #include <iostream>
 #include <algorithm>
 #include <unistd.h>
-long int get_sys_mem();
+long long get_sys_mem();
 
 int verify_dimension(int dim);
